+++ src/PodLoad.Common/Contracts/ClientReportResponse.cs
@@ -5,7 +5,7 @@ namespace PodLoad.Common.Contracts
     public class ClientReportResponse
     {
         public TimeSpan DesiredDelay { get; set; }
-        public ulong DesiredMemoryAllocated { get; set; }
+        public uint DesiredMemoryAllocated { get; set; }
         public TimeSpan KeepAliveInterval { get; set; } 
     }
 }
