+++ src/PodLoad.Server/Models/ClientInfo.cs
@@ -7,8 +7,8 @@ public class ClientInfo
     public Guid Id { get; set; }
     public string HostName { get; set; }
     public IPAddress IpAddress { get; set; }
-    public ulong MemoryAllocated { get; set; }
-    public ulong DesiredMemoryAllocated { get; set; }
+    public uint MemoryAllocated { get; set; }
+    public uint DesiredMemoryAllocated { get; set; }
     public TimeSpan Delay { get; set; }
     public TimeSpan DesiredDelay { get; set; }
     public DateTime LastUpdated { get; set; } = DateTime.Now;
